Simplify tests for variants `websockets.tutorial003`
