+++ tests/test_tutorial/test_websockets/test_tutorial003.py
@@ -1,16 +1,45 @@
+import importlib
+from types import ModuleType
+
+import pytest
 from fastapi.testclient import TestClient
 
-from docs_src.websockets.tutorial003 import app, html
-
-client = TestClient(app)
+from ...utils import needs_py39
 
 
-def test_get():
+@pytest.fixture(
+    name="mod",
+    params=[
+        pytest.param("tutorial003"),
+        pytest.param("tutorial003_py39", marks=needs_py39),
+    ],
+)
+def get_mod(request: pytest.FixtureRequest):
+    mod = importlib.import_module(f"docs_src.websockets.{request.param}")
+
+    return mod
+
+
+@pytest.fixture(name="html")
+def get_html(mod: ModuleType):
+    return mod.html
+
+
+@pytest.fixture(name="client")
+def get_client(mod: ModuleType):
+    client = TestClient(mod.app)
+
+    return client
+
+
+@needs_py39
+def test_get(client: TestClient, html: str):
     response = client.get("/")
     assert response.text == html
 
 
-def test_websocket_handle_disconnection():
+@needs_py39
+def test_websocket_handle_disconnection(client: TestClient):
     with client.websocket_connect("/ws/1234") as connection, client.websocket_connect(
         "/ws/5678"
     ) as connection_two:
